+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Improved Telemetry UI-related data collection from the main thread.
 * Reported road names will now also respect user languages preferences.
 * Fixed a UI bug when the opened steps list view was not updated and sometimes had a blank space at the top.
+* Added more feedback categories for Active Guidance.
 
 ## v3.3.0-rc.1